Refactoring

+++ CwkSocial.Application/Posts/CommandHandlers/DeletePostHandler.cs
@@ -28,32 +28,29 @@ namespace CwkSocial.Application.Posts.CommandHandlers
 
             try
             {
-                var post = await _ctx.Posts.FirstOrDefaultAsync(p => p.PostId == request.PostId);
+                var post = await _ctx.Posts.FirstOrDefaultAsync(p => p.PostId == request.PostId, cancellationToken: cancellationToken);
 
                 if (post is null)
                 {
-                    result.IsError = true;
-                    var error = new Error
-                    {
-                        Code = ErrorCode.NotFound,
-                        Message = $"No post found with ID {request.PostId}"
-                    };
-                    result.Errors.Add(error);
+                    result.AddError(ErrorCode.NotFound,
+                        string.Format(PostsErrorMessages.PostNotFound, request.PostId));
                     return result;
                 }
+
+                if(post.UserProfileId != request.UserProfileId)
+                {
+                    result.AddError(ErrorCode.PostDeleteNotPossible, PostsErrorMessages.PostDeleteNotPossible);
+                    return result;
+                }
+
                 _ctx.Posts.Remove(post);
-                await _ctx.SaveChangesAsync();
+                await _ctx.SaveChangesAsync(cancellationToken);
 
                 result.Payload = post;
             }
             catch (Exception e)
             {
-                var error = new Error
-                {
-                    Code = ErrorCode.UnknowError,
-                    Message = $"{e.Message}"};
-                result.IsError = true;
-                result.Errors.Add(error);
+                result.AddUnknownError(e.Message);
             }
 
             return result;
